fix: rename command to bookings from golf and fix linting
